Added more debug code to find problem.

+++ src/main/java/test/config/constants/signatures/paramVals/TRIAL_STATUS_ENUMS.java
@@ -48,80 +48,80 @@ public class TRIAL_STATUS_ENUMS {
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int INIT_ERROR_          = 0;
+    public static long INIT_ERROR_          = 0;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int STUB_CREATED_        = 1;
+    public static long STUB_CREATED_        = 1;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int CREATED_LINKED_      = 2;
+    public static long CREATED_LINKED_      = 2;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int KNOWNOF_             = 3;
+    public static long KNOWNOF_             = 3;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int CONFIRMED_           = 4;
+    public static long CONFIRMED_           = 4;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int PROGRESS_            = 5;
+    public static long PROGRESS_            = 5;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int EXPIRED_COMPLETED_   = 6;
+    public static long EXPIRED_COMPLETED_   = 6;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int EXPIRED_TIME_OUT_    = 7;
+    public static long EXPIRED_TIME_OUT_    = 7;
     /** "_" at end because should match instance vars.
      *  For documentation, see instance variable version. **/
     @UniqueStaticValue
-    public static int EXPIRED_COMPROMISED_ = 8;
+    public static long EXPIRED_COMPROMISED_ = 8;
     
     /** If zero, is initialization error. Integers initialize at zero.
      *  So we DONT want to use an enum of zero. As it could make error
      *  hard to find. **/
     @UniqueInstanceValue
-    public final int INIT_ERROR         = TRIAL_STATUS_ENUMS.INIT_ERROR_;
+    public final long INIT_ERROR         = TRIAL_STATUS_ENUMS.INIT_ERROR_;
     
     /** New trial has been created. But has not been linked to ANYTHING.
      *  Not a test, not a token. It is a stub. **/
     @UniqueInstanceValue
-    public final int STUB_CREATED       = TRIAL_STATUS_ENUMS.STUB_CREATED_;
+    public final long STUB_CREATED       = TRIAL_STATUS_ENUMS.STUB_CREATED_;
     
     /** New trial has been created. And the token has been linked
      *  to it. But the token has not been confirmed to be used. **/
     @UniqueInstanceValue
-    public final int CREATED_LINKED      = TRIAL_STATUS_ENUMS.CREATED_LINKED_;
+    public final long CREATED_LINKED      = TRIAL_STATUS_ENUMS.CREATED_LINKED_;
     
     /** Known of: Someone has entered the token. But has not said
      *  "YES" to the "Is this you?" confirmation page that comes after
      *  entering a token. **/
     @UniqueInstanceValue
-    public final int KNOWNOF            = TRIAL_STATUS_ENUMS.KNOWNOF_;
+    public final long KNOWNOF            = TRIAL_STATUS_ENUMS.KNOWNOF_;
     
     /** The ninja has confirmed that the token they entered has been linked
      *  to their identity. **/
     @UniqueInstanceValue
-    public final int CONFIRMED          = TRIAL_STATUS_ENUMS.CONFIRMED_;
+    public final long CONFIRMED          = TRIAL_STATUS_ENUMS.CONFIRMED_;
     
     /** The ninja has hit [START] to start the test. After being given
      *  a brief overview of what the test covers and the time allotted. **/
     @UniqueInstanceValue
-    public final int PROGRESS           = TRIAL_STATUS_ENUMS.PROGRESS_;
+    public final long PROGRESS           = TRIAL_STATUS_ENUMS.PROGRESS_;
     
     /** Token is expired, because ninja has completed trial. **/
     @UniqueInstanceValue
-    public final int EXPIRED_COMPLETED  = TRIAL_STATUS_ENUMS.EXPIRED_COMPLETED_;
+    public final long EXPIRED_COMPLETED  = TRIAL_STATUS_ENUMS.EXPIRED_COMPLETED_;
     
     /** Token is expired, because ninja has FAILED TO COMPLETE TEST WITHIN
      *  the alloted amount of time. **/
     @UniqueInstanceValue
-    public final int EXPIRED_TIME_OUT   = TRIAL_STATUS_ENUMS.EXPIRED_TIME_OUT_;
+    public final long EXPIRED_TIME_OUT   = TRIAL_STATUS_ENUMS.EXPIRED_TIME_OUT_;
     
     /** Token is expired, because someone hit "NO" when the confirmation page
      *  came up. Meaning ANOTHER ninja has access to a token that does NOT
@@ -129,7 +129,7 @@ public class TRIAL_STATUS_ENUMS {
      *  We cannot trust this token anymore. However, we need to give the ninja
      *  who was supposed to use this token another test. **/
     @UniqueInstanceValue
-    public final int EXPIRED_COMPROMISED = 
+    public final long EXPIRED_COMPROMISED = 
                                         TRIAL_STATUS_ENUMS.EXPIRED_COMPROMISED_;
     
     /*--------------------------------------------------------------------------
